requirements.txt
# ── Core Framework ──────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1

# ── LangChain Core ──────────────────────────────────────
langchain>=0.3.14,<0.4.0
langchain-core>=0.3.33,<0.4.0
langchain-community>=0.3.14,<0.4.0
langchain-text-splitters>=0.3.4,<0.4.0

# ── LLM Providers ──────────────────────────────────────
langchain-groq>=0.2.4,<0.3.0
langchain-google-genai>=2.0.8,<3.0.0
langchain-huggingface>=0.1.2,<0.2.0
together>=1.4.1,<2.0.0

# ── Vector Store ────────────────────────────────────────
langchain-pinecone>=0.2.1,<0.3.0
pinecone-client>=5.0.1,<6.0.0

# ── Embeddings ──────────────────────────────────────────
sentence-transformers>=3.3.1,<4.0.0

# ── Database + Auth ─────────────────────────────────────
supabase>=2.11.0,<3.0.0

# ── PDF Parsing ─────────────────────────────────────────
PyMuPDF>=1.25.0,<2.0.0

# ── Email ───────────────────────────────────────────────
resend>=2.7.0,<3.0.0

# ── Utilities ───────────────────────────────────────────
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
